requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
rich>=13.0.0
textual>=0.45.0
websockets>=11.0.0
plyer>=2.1.0
keyboard>=0.13.5
pyperclip>=1.8.2
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.8.0
aiolimiter>=1.1.0
pyyaml>=6.0
jinja2>=3.1.0
structlog>=23.0.0
python-multipart>=0.0.20
